refactor: move ScorePie into Score/
